+++ pkg/middleware/parallel/parallel.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/cloudwego/hertz/pkg/app"
 	"github.com/go-viper/mapstructure/v2"
-	"github.com/nite-coder/bifrost/internal/pkg/task"
+	"github.com/nite-coder/bifrost/internal/pkg/safety"
 	"github.com/nite-coder/bifrost/pkg/config"
 	"github.com/nite-coder/bifrost/pkg/middleware"
 )
@@ -34,7 +34,7 @@ func (m *ParallelMiddleware) ServeHTTP(ctx context.Context, c *app.RequestContex
 	waitGroup.Add(len(m.options))
 
 	for _, option := range m.options {
-		go task.Runner(ctx, func() {
+		go safety.Go(ctx, func() {
 			defer func() {
 				if r := recover(); r != nil {
 					// convert panic to an error